% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{cronbach2}
\alias{cronbach2}
\title{cronbach2}
\usage{
cronbach2(
  D,
  Key = scoring.key("NEO EE"),
  Mid = 2,
  what = "DF",
  format = "m",
  print.all = FALSE
)
}
\arguments{
\item{D}{x}

\item{Key}{x}

\item{Mid}{x}

\item{what}{x}

\item{format}{'m' or 's'}

\item{print.all}{FALSE}
}
\value{
x
}
\description{
cronbach2
}
\details{
x
}
