\name{ind_fun_pemom}
\alias{ind_fun_pemom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
the log-marginal likelhood function based on peMoM priors and inverse gamma prior (0.01,0.01)
}
\description{
the log-marginal likelhood value of a model, based on the  peMoM prior on the regression coefficients  and inverse gamma prior (0.01,0.01) on the variance.}
\usage{
ind_fun_pemom(ind2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ind2}{
   the index set of variables in a model}
}
\seealso{
\code{\link{ind_fun_g}}, \code{\link{ind_fun_pimom}}
}
