\name{ind_fun_pimom}
\alias{ind_fun_pimom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
the log-marginal likelhood function based on piMoM priors
}
\description{
the log-marginal likelhood value of a model, based on the  piMoM prior on the regression coefficients and inverse gamma prior (0.01,0.01) on the variance.}
\usage{
ind_fun_pimom(ind2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ind2}{
   the index set of variables in a model}
}
\seealso{
\code{\link{ind_fun_g}}, \code{\link{ind_fun_pemom}}
}
