% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5.R
\docType{package}
\name{s5-package}
\alias{s5-package}
\alias{s5}
\title{s5  package}
\description{
\tabular{ll}{
Package: \tab psycho\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2021-05-09\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\details{
S5 functions
}
\examples{
NULL


}
\seealso{
\code{\link{NULL}}
}
\author{
Kenn Konstabel \email{lebatsnok AT gmail.com}
}
\keyword{package}
