% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score}
\alias{score}
\title{Score}
\usage{
score(
  D,
  K,
  R = NULL,
  Mid = 2,
  S = NULL,
  na.action = "exclude",
  na.tol = 0,
  hofstee = FALSE,
  lims = c(0, 4),
  average = FALSE
)
}
\arguments{
\item{D}{a df}

\item{K}{a key}

\item{R}{something}

\item{Mid}{middle value}

\item{S}{sth}

\item{na.action}{exclude}

\item{na.tol}{na.tol}

\item{hofstee}{boolean}

\item{lims}{lims}

\item{average}{boolena}
}
\value{
x
}
\description{
Score
}
\details{
x
}
