% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score2}
\alias{score2}
\title{Score2}
\usage{
score2(
  D,
  test = "NEO EE",
  Mid = 2,
  lims = NULL,
  na.action = "exclude",
  na.tol1 = 1,
  na.tol2 = 3,
  what = "DF",
  override.mid = FALSE,
  ...
)
}
\arguments{
\item{D}{df}

\item{test}{test}

\item{Mid}{mid}

\item{lims}{lims}

\item{na.action}{na.action}

\item{na.tol1}{na.tol1}

\item{na.tol2}{na.tol2}

\item{what}{what}

\item{override.mid}{sth}

\item{...}{dot dot dot}
}
\value{
x
}
\description{
Score2
}
\details{
x
}
