% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{sdi}
\alias{sdi}
\title{sdi}
\usage{
sdi(andmed, soso, hofstee = FALSE, lims = c(0, 4, 1, 7))
}
\arguments{
\item{andmed}{x}

\item{soso}{x}

\item{hofstee}{x}

\item{lims}{x}
}
\value{
x
}
\description{
sdi
}
\details{
x
}
