% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{sdi2}
\alias{sdi2}
\title{sdi2}
\usage{
sdi2(D, key = "NEO EE", na.tol = 10, hofstee = TRUE, lims = c(0, 4, 1, 7))
}
\arguments{
\item{D}{x}

\item{key}{x}

\item{na.tol}{x}

\item{hofstee}{x}

\item{lims}{x}
}
\value{
x
}
\description{
sdi2
}
\details{
x
}
