% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{cronbach2}
\alias{cronbach2}
\title{cronbach2}
\usage{
cronbach2(
  D,
  Key = scoring.key("NEO EE"),
  Mid = 2,
  what = "DF",
  format = "m",
  print.all = FALSE
)
}
\arguments{
\item{D}{data set}

\item{Key}{scoring key}

\item{Mid}{scale middle point}

\item{what}{with hierarchical scoring keys, use "D" to compute 'domain' scores , "F" for 'facet' scores,  and "DF" for both.}

\item{format}{'m' or 's'}

\item{print.all}{FALSE}
}
\value{
x
}
\description{
cronbach2
}
\details{
x
}
